\name{shortest.paths}
\alias{shortest.paths}
\alias{.shortest.paths}
\alias{.walk.paths}
\alias{.skim.paths}
\alias{.load.paths}
\alias{.intercept.paths}

\title{
Low-Level Functions for Shortest Path Manipulation
}
\description{
These functions for manipulating shortest paths are used to build
higher-level algorithms, such as for highway assignment, computing
network skims (applying a function across a path), and loading an
origin-destination matrix onto the links on the shortest path between
each origin-destination pair.
}
\usage{
.shortest.paths(network.set, costs)
.load.paths(paths, demand)
.skim.paths(paths, costs, empty.value = 0)
.intercept.paths(paths, links)
.walk.paths(paths, origins, dests, permute=TRUE)
}
\arguments{
  \item{network.set}{
  Network descriptor (generated by \code{\link{build.network.set}})
}
  \item{demand}{
  A matrix of demands between each origin/destination pair
}
  \item{costs}{
  A vector of costs for each link in the network
}
  \item{paths}{
  A shortest-path tree for a particular network
}
  \item{empty.value}{
  The value to be used for origin/destination pairs that have no path, and
  also for intrazonal (diagonal) cells
}
  \item{links}{
  A vector selecting links of interest in the network; the intercept matrix will contain true for
  any origin/destination pair whose shortest path crosses one of these links.
}
  \item{origins}{
  A vector of origin zones
}
  \item{dests}{
  A vector of destination zones
}
  \item{permute}{
  If true, returns paths for all combinations of \code{origins} and \code{dests}; otherwise, \code{origins} and \code{dests} must be vectors of the same length and paths will only be built for each corresponding origin and destination
}
}
\details{
Details yet to come.
}
\value{
Each function returns something a bit different.
}
\author{
Jeremy Raw
}

\seealso{
\code{\link{highway.assign}}
}
\examples{
\dontrun{print("Examples yet to come.")}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
