\name{package-travelr}
\alias{package-travelr}
\alias{travelr}

\title{
TravelR: Tools for writing travel demand models
}
\description{
TravelR provides some common, specialized tools for travel demand modeling in R
}
\details{

The \pkg{travelr} package provides tools for travel demand modeling that are not otherwise available in \R.

In particular, the package includes a set of functions for doing highway assignment, and an iterative proportional
factoring routine for distributing marginal adjustments of an array across the elements of the array (commonly referred
to as a \\dQuote{Fratar} function after).  Additional tools may appear later, most notably tools for transit
path-building and assignment.

\pkg{travelr} is intended both for producing \\dQuote{working models} and for researching model algorithms, particularly
algorithms for highway assignment.

}
